New post data structure, stick/locked threads, bugs.
Wrote a new posting data structure that'll have much better longevity;
as the current one would start to be problematic after a thread had a
few thousand posts or so. Limitation should be closer to simply what
mysql/your disk space can handle now. This is not backwards compatible
with the old data structure, old installations will not be able to be
upgraded to this version; mostly because I'm lazy and won't write
something to convert the format. But who even uses this forum anyways?

Added controls so that admins and users can lock/unlock threads. Admins
also have a control to sticky/unsticky threads. Added a visual indicator
for sticky and locked threads.

Some small aesthetic adjustments and bugfixes as usual.