Refactored code and added pages to thread index.
Refactored mysql code such that only one mysql connection is used for the entire script. This reduces page generation time since the script doesn't reconnect to the database for each individual query. This means there is also a 'in X queries' stat at the bottom of the page now.

Also added pages to the thread index to allow users to browse multiple pages of threads. Not sure how I overlooked this until now.